+++ Runtime/Scripts/IObservable.cs
@@ -1,0 +1,7 @@
+namespace NarrationsJouables
+{
+    public interface IObservable
+    {
+        public bool ObservationStateChanged(bool _observed, float distance = -1f);
+    }
+}