+++ CHANGELOG.md
@@ -4,11 +4,14 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.0.2] - 2026-04-14
 
 ### Added
 
 - Title screen template scene, prefab and scripts
+- Interaction scripts and prefabs
+- Highlight effect shader
+- Sample scene with interaction examples
 
 ## [1.0.1] - 2026-03-31
 
@@ -16,7 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Fixed version number in package.json
 
-## [1.0.0] - 2026-03-31
+## [1.0.0] - 2026-03-31
 
 ### Added